Enables array based bounding boxes
