+++ node-hub/dora-rerun/src/lib.rs
@@ -3,18 +3,15 @@
 use std::{collections::HashMap, env::VarError};
 
 use dora_node_api::{
-    arrow::{
-        array::{AsArray, Float32Array, Float64Array, StringArray, StructArray, UInt8Array},
-        datatypes::Float32Type,
-    },
+    arrow::array::{Array, Float32Array, Float64Array, StringArray, UInt8Array},
     DoraNode, Event, Parameter,
 };
 use eyre::{eyre, Context, ContextCompat, Result};
 
 use rerun::{
     components::ImageBuffer, external::re_types::ArrowBuffer, ImageFormat, Points3D, SpawnOptions,
-    Text,
 };
+pub mod boxes2d;
 pub mod series;
 pub mod urdf;
 use series::update_series;
@@ -187,75 +184,7 @@ pub fn lib_main() -> Result<()> {
                     }
                 })?;
             } else if id.as_str().contains("boxes2d") {
-                let bbox_struct: StructArray = data.to_data().into();
-                let format =
-                    if let Some(Parameter::String(format)) = metadata.parameters.get("format") {
-                        format
-                    } else {
-                        "xyxy"
-                    };
-
-                // Cast Bbox
-                let bbox_buffer = bbox_struct
-                    .column_by_name("bbox")
-                    .context("Did not find labels field within bbox struct")?;
-                let bbox = bbox_buffer
-                    .as_list_opt::<i32>()
-                    .context("Could not deserialize bbox as list")?
-                    .values();
-                let bbox = bbox
-                    .as_primitive_opt::<Float32Type>()
-                    .context("Could not get bbox value as list")?
-                    .values();
-
-                // Cast Labels
-                let labels_buffer = bbox_struct
-                    .column_by_name("labels")
-                    .context("Did not find labels field within bbox struct")?;
-                let labels = labels_buffer
-                    .as_list_opt::<i32>()
-                    .context("Could not deserialize labels as list")?
-                    .values();
-                let labels = labels
-                    .as_string_opt::<i32>()
-                    .context("Could not deserialize labels as string")?;
-                let labels: Vec<Text> = labels.iter().map(|x| Text::from(x.unwrap())).collect();
-
-                // Cast confidence
-                let conf_buffer = bbox_struct
-                    .column_by_name("conf")
-                    .context("Did not find conf field within bbox struct")?;
-                let conf = conf_buffer
-                    .as_list_opt::<i32>()
-                    .context("Could not deserialize conf as list")?
-                    .values();
-                let _conf = conf
-                    .as_primitive_opt::<Float32Type>()
-                    .context("Could not deserialize conf as string")?;
-
-                let mut centers = vec![];
-                let mut sizes = vec![];
-
-                if format == "xywh" {
-                    bbox.chunks(4).for_each(|block| {
-                        if let [x, y, w, h] = block {
-                            centers.push((*x, *y));
-                            sizes.push((*w, *h));
-                        }
-                    });
-                } else if format == "xyxy" {
-                    bbox.chunks(4).for_each(|block| {
-                        if let [min_x, min_y, max_x, max_y] = block {
-                            centers.push(((max_x + min_x) / 2., (max_y + min_y) / 2.));
-                            sizes.push(((max_x - min_x), (max_y - min_y)));
-                        }
-                    });
-                }
-                rec.log(
-                    id.as_str(),
-                    &rerun::Boxes2D::from_centers_and_sizes(centers, sizes).with_labels(labels),
-                )
-                .wrap_err("Could not log Boxes2D")?;
+                boxes2d::update_boxes2d(&rec, id, data, metadata).context("update boxes 2d")?;
             } else if id.as_str().contains("jointstate") {
                 let buffer: &Float32Array = data
                     .as_any()
